make enum to entity cleanner
