+++ office-space-backend/src/main/java/com/ourfancyteamname/officespace/security/payload/UserDetailsPrinciple.java
@@ -2,7 +2,7 @@ package com.ourfancyteamname.officespace.security.payload;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.ourfancyteamname.officespace.dtos.security.RoleDto;
-import com.ourfancyteamname.officespace.postgres.enums.PermissionCode;
+import com.ourfancyteamname.officespace.enums.PermissionCode;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
