+++ office-space-model/src/main/java/com/ourfancyteamname/officespace/enums/Gender.java
@@ -1,0 +1,10 @@
+package com.ourfancyteamname.officespace.enums;
+
+public enum Gender implements PersistableEnum<String> {
+  MALE, FEMALE;
+
+  @Override
+  public String getName() {
+    return this.name();
+  }
+}
