+++ office-space-model/src/main/java/com/ourfancyteamname/officespace/enums/PermissionCode.java
@@ -1,0 +1,10 @@
+package com.ourfancyteamname.officespace.enums;
+
+public enum PermissionCode implements PersistableEnum<String> {
+  DELETE_USER;
+
+  @Override
+  public String getName() {
+    return this.name();
+  }
+}
