+++ office-space-model/src/main/java/com/ourfancyteamname/officespace/enums/PersistableEnum.java
@@ -1,0 +1,5 @@
+package com.ourfancyteamname.officespace.enums;
+
+public interface PersistableEnum<T> {
+  T getName();
+}
