+++ office-space-postgres/src/main/java/com/ourfancyteamname/officespace/postgres/converters/PermissionCodeConverter.java
@@ -1,0 +1,12 @@
+package com.ourfancyteamname.officespace.postgres.converters;
+
+import com.ourfancyteamname.officespace.enums.PermissionCode;
+
+import javax.persistence.Converter;
+
+@Converter(autoApply = true)
+public class PermissionCodeConverter extends AbstractHibernateEnumConverter<PermissionCode, String> {
+  public PermissionCodeConverter() {
+    super(PermissionCode.class);
+  }
+}
