--- office-space-postgres/src/main/java/com/ourfancyteamname/officespace/postgres/enums/PermissionCode.java
@@ -1,18 +1,0 @@
-package com.ourfancyteamname.officespace.postgres.enums;
-
-import com.ourfancyteamname.officespace.postgres.converters.AbstractHibernateEnumConverter;
-
-public enum PermissionCode implements PersistableEnum<String> {
-  DELETE_USER;
-
-  @Override
-  public String getName() {
-    return this.name();
-  }
-
-  public static class HibernateConverter extends AbstractHibernateEnumConverter<PermissionCode, String> {
-    public HibernateConverter() {
-      super(PermissionCode.class);
-    }
-  }
-}
