--- office-space-postgres/src/main/java/com/ourfancyteamname/officespace/postgres/enums/PersistableEnum.java
@@ -1,5 +1,0 @@
-package com.ourfancyteamname.officespace.postgres.enums;
-
-public interface PersistableEnum<T> {
-  T getName();
-}
