+++ office-space-postgres/src/main/java/com/ourfancyteamname/officespace/postgres/repos/PermissionRepository.java
@@ -1,7 +1,7 @@
 package com.ourfancyteamname.officespace.postgres.repos;
 
+import com.ourfancyteamname.officespace.enums.PermissionCode;
 import com.ourfancyteamname.officespace.postgres.entities.Permission;
-import com.ourfancyteamname.officespace.postgres.enums.PermissionCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
